requirements.txt
Flask==2.3.4
google-cloud-storage==2.11.0
google-cloud-pubsub==2.16.0
